perf(garmin): offload webhook processing to background tasks
