+++ backend/app/integrations/celery/core.py
@@ -71,6 +71,8 @@ def create_celery() -> Celery:
         },
         task_routes={
             "app.integrations.celery.tasks.process_sdk_upload_task.process_sdk_upload": {"queue": "sdk_sync"},
+            "app.integrations.celery.tasks.garmin_webhook_task.process_garmin_ping": {"queue": "garmin_sync"},
+            "app.integrations.celery.tasks.garmin_webhook_task.process_garmin_push": {"queue": "garmin_sync"},
         },
     )
