+++ backend/scripts/start/worker.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 set -e -x
 
-uv run celery -A app.main:celery_app worker --loglevel=info --pool=threads -Q default,sdk_sync
+uv run celery -A app.main:celery_app worker --loglevel=info --pool=threads -Q default,sdk_sync,garmin_sync
