+++ backend/tests/conftest.py
@@ -169,8 +169,8 @@ def mock_celery_tasks(monkeypatch: pytest.MonkeyPatch) -> Generator[MagicMock, N
 
     with (
         patch("celery.current_app") as mock_celery,
-        # Patch Garmin backfill task dispatched from webhook route
-        patch("app.api.routes.v1.garmin_webhooks.trigger_next_pending_type", mock_task),
+        # Patch Garmin backfill chaining task used by webhook processing tasks
+        patch("app.integrations.celery.tasks.garmin_webhook_task.trigger_next_pending_type", mock_task),
     ):
         # Configure Celery to use in-memory broker and result backend
         # We Mock the conf object to return our test settings
@@ -212,8 +212,6 @@ def mock_external_apis() -> Generator[dict[str, MagicMock], None, None]:
     mock_s3.head_bucket.return_value = {}
     mock_s3.put_object.return_value = {"ETag": "test-etag"}
 
-    webhook_module = "app.api.routes.v1.garmin_webhooks"
-
     with (
         patch("httpx.AsyncClient") as mock_httpx,
         patch("boto3.client", return_value=mock_s3) as mock_boto3,
@@ -226,13 +224,6 @@ def mock_external_apis() -> Generator[dict[str, MagicMock], None, None]:
         patch(
             "app.services.apple.apple_xml.presigned_url_service.presigned_url_service.s3_client", mock_s3, create=True
         ),
-        patch(f"{webhook_module}.get_trace_id", return_value=None),
-        patch(f"{webhook_module}.mark_type_success", return_value=False),
-        patch(
-            f"{webhook_module}.get_backfill_status",
-            return_value={"overall_status": "complete", "current_window": 0, "total_windows": 0},
-        ),
-        patch(f"{webhook_module}.trigger_next_pending_type", return_value={}),
     ):
         mocks["httpx"] = mock_httpx
         mocks["boto3"] = mock_boto3
